Move ShimLoader to API
