+++ sql-plugin-api/src/main/scala/com/nvidia/spark/rapids/package.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.rapids.execution
+package com.nvidia.spark.rapids
 
-object UnshimmedTrampolineUtil {
-  def sparkClassLoader: ClassLoader = {
-    org.apache.spark.util.Utils.getContextOrSparkClassLoader
-  }
-
-  def classIsLoadable(className: String): Boolean =
-    org.apache.spark.util.Utils.classIsLoadable(className)
+/**
+ * Trait that makes it easy to check whether we are dealing with the
+ * a RAPIDS Shuffle Manager
+ */
+trait RapidsShuffleManagerLike {
+  def isDriver: Boolean
+  def initialize: Unit
 }
