+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/ExplainPlan.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION.
+ * Copyright (c) 2021-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -62,7 +62,7 @@ object ExplainPlan {
   @throws[IllegalStateException]
   def explainPotentialGpuPlan(df: DataFrame, explain: String = "ALL"): String = {
     try {
-      ShimLoader.newExplainPlan.explainPotentialGpuPlan(df, explain)
+      ShimLoaderTemp.newExplainPlan.explainPotentialGpuPlan(df, explain)
     } catch {
       case ia: IllegalArgumentException => throw ia
       case is: IllegalStateException => throw is
