+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/delta/DeltaProvider.scala
@@ -16,7 +16,7 @@
 
 package com.nvidia.spark.rapids.delta
 
-import com.nvidia.spark.rapids.{CreatableRelationProviderRule, ExecRule, RunnableCommandRule, ShimLoader, SparkPlanMeta}
+import com.nvidia.spark.rapids.{CreatableRelationProviderRule, ExecRule, RunnableCommandRule, ShimLoaderTemp, SparkPlanMeta}
 
 import org.apache.spark.sql.Strategy
 import org.apache.spark.sql.execution.{FileSourceScanExec, SparkPlan}
@@ -50,7 +50,7 @@ trait DeltaProvider {
 
 object DeltaProvider {
   private lazy val provider = {
-    ShimLoader.newDeltaProbe().getDeltaProvider
+    ShimLoaderTemp.newDeltaProbe().getDeltaProvider
   }
 
   def apply(): DeltaProvider = provider
