+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/RapidsShuffleInternalManagerBase.scala
@@ -1421,72 +1421,3 @@ class RapidsShuffleInternalManagerBase(conf: SparkConf, val isDriver: Boolean)
     }
   }
 }
-
-/**
- * Trait that makes it easy to check whether we are dealing with the
- * a RAPIDS Shuffle Manager
- */
-trait RapidsShuffleManagerLike {
-  def isDriver: Boolean
-  def initialize: Unit
-}
-
-/**
- * A simple proxy wrapper allowing to delay loading of the
- * real implementation to a later point when ShimLoader
- * has already updated Spark classloaders.
- *
- * @param conf
- * @param isDriver
- */
-class ProxyRapidsShuffleInternalManagerBase(
-    conf: SparkConf,
-    override val isDriver: Boolean
-) extends RapidsShuffleManagerLike with Proxy with ShuffleManager {
-
-  // touched in the plugin code after the shim initialization
-  // is complete
-  lazy val self: ShuffleManager = ShimLoader.newInternalShuffleManager(conf, isDriver)
-      .asInstanceOf[ShuffleManager]
-
-  // This function touches the lazy val `self` so we actually instantiate
-  // the manager. This is called from both the driver and executor.
-  // In the driver, it's mostly to display information on how to enable/disable the manager,
-  // in the executor, the UCXShuffleTransport starts and allocates memory at this time.
-  override def initialize: Unit = self
-
-  def getWriter[K, V](
-      handle: ShuffleHandle,
-      mapId: Long,
-      context: TaskContext,
-      metrics: ShuffleWriteMetricsReporter
-  ): ShuffleWriter[K, V] = {
-    self.getWriter(handle, mapId, context, metrics)
-  }
-
-  def getReader[K, C](
-      handle: ShuffleHandle,
-      startMapIndex: Int,
-      endMapIndex: Int,
-      startPartition: Int,
-      endPartition: Int,
-      context: TaskContext,
-      metrics: ShuffleReadMetricsReporter): ShuffleReader[K, C] = {
-    self.getReader(handle,
-      startMapIndex, endMapIndex, startPartition, endPartition,
-      context, metrics)
-  }
-
-  def registerShuffle[K, V, C](
-      shuffleId: Int,
-      dependency: ShuffleDependency[K, V, C]
-  ): ShuffleHandle = {
-    self.registerShuffle(shuffleId, dependency)
-  }
-
-  def unregisterShuffle(shuffleId: Int): Boolean = self.unregisterShuffle(shuffleId)
-
-  def shuffleBlockResolver: ShuffleBlockResolver = self.shuffleBlockResolver
-
-  def stop(): Unit = self.stop()
-}
